Consolidate all the tests together
